Light bg: use DarkGreen instead of DarkMagenta
I had originally based the colors on xterm colors and the default green
on a white background is hard to read but for consistency I want to
leave it as DarkGreen to match the Green of the dark background. If I
use a white background in xterm or urxvt I think I would also want to
darken the green color anyway so I can read diff outputs, etc.